+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const defaultIndentation = 16;
+export const defaultDropIndicatorOffset = 0;
